requirements.txt
asgiref==3.8.1
black==24.4.2
bleach==6.1.0
click==8.1.7
colorama==0.4.6
Django==4.2.13
django-admin-autocomplete-filter==0.7.1
django-cors-headers==4.3.1
django-filter==24.2
django-simple-history==3.7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-api-logger==1.1.15
gunicorn==22.0.0
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing-extensions==4.12.1
tzdata==2024.1
webencodings==0.5.1
